Restructure code around and add redirect template
