+++ src/redirect.rs
@@ -1,10 +1,7 @@
 use crate::i18n::SUPPORTED_LOCALES;
-use rocket::http::uri::Path;
-use rocket::response::Redirect;
+use crate::render::RenderCtx;
 
-static PAGE_REDIRECTS: &[(&str, &str)] = &[
-    // Migrate pre-2018 locale for the index page
-    ("", ""),
+pub static PAGE_REDIRECTS: &[(&str, &str)] = &[
     // Pre-2018 website pages
     ("community.html", "community"),
     ("conduct.html", "policies/code-of-conduct"),
@@ -47,7 +44,7 @@ static PAGE_REDIRECTS: &[(&str, &str)] = &[
     ("governance/wgs", "governance#working-groups"),
 ];
 
-static STATIC_FILES_REDIRECTS: &[(&str, &str)] = &[
+pub static STATIC_FILES_REDIRECTS: &[(&str, &str)] = &[
     // Pre-2018 whitepaper locations
     (
         "pdfs/Rust-npm-Whitepaper.pdf",
@@ -84,52 +81,56 @@ static PRE_2018_LOCALES: &[&str] = &[
     "ru-RU", "sv-SE", "vi-VN",
 ];
 
-pub(crate) fn maybe_redirect(path: Path) -> Option<Redirect> {
-    let path = path.segments().collect::<Vec<_>>().join("/");
+pub fn create_redirects(ctx: &RenderCtx) -> anyhow::Result<()> {
+    // Static file redirects, no need to support languages
+    // for (src, dst) in STATIC_FILES_REDIRECTS {
+    //     render_redirect(ctx, src, dst)?;
+    // }
 
-    if let Some((_, dest)) = STATIC_FILES_REDIRECTS.iter().find(|(src, _)| src == &path) {
-        return Some(Redirect::permanent(*dest));
-    }
-
-    let (locale, path) = if let Some((first, rest)) = path.split_once('/') {
-        if SUPPORTED_LOCALES.contains(&first) {
-            (Locale::Present(first), rest)
-        // After the 2018 website redesign some of the locales were removed and some were
-        // renamed (removing the country code). This handles both cases, either calculating the
-        // renamed locale or marking the locale as "specified but missing", which triggers a
-        // temporary redirect instead of a permanent redirect.
-        } else if PRE_2018_LOCALES.contains(&first) {
-            if let Some(locale) = convert_locale_from_pre_2018(first) {
-                (Locale::Present(locale), rest)
-            } else {
-                (Locale::SpecifiedButMissing, rest)
-            }
-        } else {
-            (Locale::NotSpecified, path.as_str())
-        }
-    } else if PRE_2018_LOCALES.contains(&path.as_str()) {
-        // If the whole path is a pre-2018 locale handle it as a localized index page.
-        if let Some(locale) = convert_locale_from_pre_2018(&path) {
-            (Locale::Present(locale), "")
-        } else {
-            (Locale::SpecifiedButMissing, "")
-        }
-    } else {
-        (Locale::NotSpecified, path.as_str())
-    };
-
-    if let Some((_, dest)) = EXTERNAL_REDIRECTS.iter().find(|(src, _)| *src == path) {
-        Some(Redirect::permanent(*dest))
-    } else if let Some((_, dest)) = PAGE_REDIRECTS.iter().find(|(src, _)| *src == path) {
-        let dest = format!("/{dest}");
-        match locale {
-            Locale::Present("en-US") | Locale::NotSpecified => Some(Redirect::permanent(dest)),
-            Locale::Present(locale) => Some(Redirect::permanent(format!("/{locale}{dest}"))),
-            Locale::SpecifiedButMissing => Some(Redirect::temporary(dest)),
-        }
-    } else {
-        None
-    }
+    // if let Some((_, dest)) = STATIC_FILES_REDIRECTS.iter().find(|(src, _)| src == &path) {
+    //     return Some(Redirect::permanent(*dest));
+    // }
+    //
+    // let (locale, path) = if let Some((first, rest)) = path.split_once('/') {
+    //     if SUPPORTED_LOCALES.contains(&first) {
+    //         (Locale::Present(first), rest)
+    //     // After the 2018 website redesign some of the locales were removed and some were
+    //     // renamed (removing the country code). This handles both cases, either calculating the
+    //     // renamed locale or marking the locale as "specified but missing", which triggers a
+    //     // temporary redirect instead of a permanent redirect.
+    //     } else if PRE_2018_LOCALES.contains(&first) {
+    //         if let Some(locale) = convert_locale_from_pre_2018(first) {
+    //             (Locale::Present(locale), rest)
+    //         } else {
+    //             (Locale::SpecifiedButMissing, rest)
+    //         }
+    //     } else {
+    //         (Locale::NotSpecified, path.as_str())
+    //     }
+    // } else if PRE_2018_LOCALES.contains(&path.as_str()) {
+    //     // If the whole path is a pre-2018 locale handle it as a localized index page.
+    //     if let Some(locale) = convert_locale_from_pre_2018(&path) {
+    //         (Locale::Present(locale), "")
+    //     } else {
+    //         (Locale::SpecifiedButMissing, "")
+    //     }
+    // } else {
+    //     (Locale::NotSpecified, path.as_str())
+    // };
+    //
+    // if let Some((_, dest)) = EXTERNAL_REDIRECTS.iter().find(|(src, _)| *src == path) {
+    //     Some(Redirect::permanent(*dest))
+    // } else if let Some((_, dest)) = PAGE_REDIRECTS.iter().find(|(src, _)| *src == path) {
+    //     let dest = format!("/{dest}");
+    //     match locale {
+    //         Locale::Present("en-US") | Locale::NotSpecified => Some(Redirect::permanent(dest)),
+    //         Locale::Present(locale) => Some(Redirect::permanent(format!("/{locale}{dest}"))),
+    //         Locale::SpecifiedButMissing => Some(Redirect::temporary(dest)),
+    //     }
+    // } else {
+    //     None
+    // }
+    Ok(())
 }
 
 fn convert_locale_from_pre_2018(pre_2018: &str) -> Option<&str> {
