+++ templates/redirect.html.hbs
@@ -1,0 +1,6 @@
+<!doctype html>
+<html>
+  <meta http-equiv="refresh" content="0; url={{ data.url }}">
+  <body>Please go to <a href="{{ data.url }}">{{ data.url }}</a></body>
+  <script>location.href = "{{ data.url }}";</script>
+</html>
